feat: merged tests from test_deduplication_lifo.py and test_deduplication_lifo_with_nullifiers.py into 1 file
